+++ src/client/kotlin/tech/thatgravyboat/skyblockapi/api/area/hub/BazaarAPI.kt
@@ -3,7 +3,7 @@ package tech.thatgravyboat.skyblockapi.api.area.hub
 import org.jetbrains.annotations.ApiStatus
 import tech.thatgravyboat.skyblockapi.api.remote.pricing.BazaarAPI
 
-@Deprecated("Moved to remote.pricing.BazaarAPI")
+@Deprecated("Moved to remote.hypixel.pricing.BazaarAPI")
 @ApiStatus.ScheduledForRemoval(inVersion = "1.21.6 or 1.22")
 object BazaarAPI {
