+++ src/client/kotlin/tech/thatgravyboat/skyblockapi/api/area/hub/LowestBinAPI.kt
@@ -3,7 +3,7 @@ package tech.thatgravyboat.skyblockapi.api.area.hub
 import org.jetbrains.annotations.ApiStatus
 import tech.thatgravyboat.skyblockapi.api.remote.pricing.LowestBinAPI
 
-@Deprecated("Moved to remote.pricing.LowestBinAPI")
+@Deprecated("Moved to remote.hypixel.pricing.LowestBinAPI")
 @ApiStatus.ScheduledForRemoval(inVersion = "1.21.6 or 1.22")
 object LowestBinAPI {
     /** Hypixel Item Id to Prices */
