+++ src/client/kotlin/tech/thatgravyboat/skyblockapi/api/item/calculator/Pricing.kt
@@ -4,7 +4,7 @@ import org.jetbrains.annotations.ApiStatus
 import tech.thatgravyboat.skyblockapi.api.remote.pricing.BazaarAPI
 import tech.thatgravyboat.skyblockapi.api.remote.pricing.LowestBinAPI
 
-@Deprecated("Moved", ReplaceWith("remote.pricing.Pricing.getPrice(id)"))
+@Deprecated("Moved", ReplaceWith("remote.hypixel.pricing.Pricing.getPrice(id)"))
 @ApiStatus.ScheduledForRemoval(inVersion = "1.21.6 or 1.22")
 object Pricing {
     fun getPrice(id: String?): Long {
