+++ src/client/kotlin/tech/thatgravyboat/skyblockapi/api/remote/hypixel/pricing/Pricing.kt
@@ -1,9 +1,5 @@
-package tech.thatgravyboat.skyblockapi.api.remote.pricing
+package tech.thatgravyboat.skyblockapi.api.remote.hypixel.pricing
 
-import org.jetbrains.annotations.ApiStatus
-
-@Deprecated("Moved", ReplaceWith("remote.hypixel.pricing.Pricing.getPrice(id)"))
-@ApiStatus.ScheduledForRemoval(inVersion = "1.21.6 or 1.22")
 object Pricing {
     fun getPrice(id: String?): Long {
         val product = BazaarAPI.getProduct(id)
