+++ src/client/kotlin/tech/thatgravyboat/skyblockapi/api/remote/pricing/LowestBinAPI.kt
@@ -1,46 +1,26 @@
 package tech.thatgravyboat.skyblockapi.api.remote.pricing
 
-import com.google.gson.JsonObject
-import tech.thatgravyboat.skyblockapi.modules.Module
-import tech.thatgravyboat.skyblockapi.utils.Scheduling
-import tech.thatgravyboat.skyblockapi.utils.extentions.asDouble
-import tech.thatgravyboat.skyblockapi.utils.extentions.asLong
-import tech.thatgravyboat.skyblockapi.utils.extentions.asMap
-import tech.thatgravyboat.skyblockapi.utils.http.Http
-import kotlin.time.Duration.Companion.hours
-import kotlin.time.Duration.Companion.seconds
+import org.jetbrains.annotations.ApiStatus
+import tech.thatgravyboat.skyblockapi.api.remote.hypixel.pricing.LowestBinAPI
 
-private const val URL = "https://skyblock-pv.thatgravyboat.tech/auctions"
-
-@Module
+@Deprecated("Moved to remote.hypixel.pricing.LowestBinAPI")
+@ApiStatus.ScheduledForRemoval(inVersion = "1.21.6 or 1.22")
 object LowestBinAPI {
     /** Hypixel Item Id to Prices */
-    var items = mapOf<String, AuctionItem>()
-        private set
+    val items get() = LowestBinAPI.items
 
-    fun getPrice(id: String?): AuctionItem? = items.entries.find { it.key.equals(id, ignoreCase = true) }?.value
-    fun getLowestPrice(id: String?): Long? = getPrice(id)?.lowest
+    fun getPrice(id: String?): AuctionItem? {
+        val item = LowestBinAPI.getPrice(id) ?: return null
 
-    init {
-        Scheduling.schedule(0.seconds, 2.hours) {
-            fetch()
-        }
+        return AuctionItem(
+            lowest = item.lowest,
+            highest = item.highest,
+            median = item.median,
+            mean = item.mean,
+        )
     }
 
-    private suspend fun fetch() {
-        Http.getResult<JsonObject>(URL).let { res ->
-            val response = res.getOrNull() ?: return
-            items = response.asMap { id, item ->
-                val obj = item.asJsonObject
-                id to AuctionItem(
-                    obj["lowest"].asLong(0),
-                    obj["highest"].asLong(0),
-                    obj["median"].asLong(0),
-                    obj["mean"].asDouble(0.0),
-                )
-            }
-        }
-    }
+    fun getLowestPrice(id: String?): Long? = LowestBinAPI.getLowestPrice(id)
 
     data class AuctionItem(
         val lowest: Long,
